Add compatiblity with Android 9
On Android 9, TimeLapse was unable to connect to cameras because
they are accessed using plain (unencrypted) HTTP. The specific
IP addresses we need to connect to are now listed in
`app/src/main/res/xml/network_security_config.xml`.

A related change is the need to explicity list the
org.apache.http.legacy library in the manifest as
used by the app.

Finally, the application needs to ask for the new FOREGROUND_SERVICE
permission in order to work.